Move optimizer to common folder
